deduplicate index entry points
